+++ sdk/src/error.rs
@@ -17,6 +17,7 @@ use thiserror::Error;
 
 /// `Error` enumerates errors returned by most C2PA toolkit operations.
 #[derive(Debug, Error)]
+#[non_exhaustive]
 pub enum Error {
     // --- c2pa errors ---
     /// Could not find a claim with this label.
@@ -200,6 +201,9 @@ pub enum Error {
 
     #[error("file not found: {0}")]
     FileNotFound(String),
+
+    #[error("resource not found: {0}")]
+    ResourceNotFound(String),
 
     #[error("XMP read error")]
     XmpReadError,